Conflict in tests/test_brachistochrone.py::test_backend:
  - HEAD added a 3-axis parametrize (backend × constraint_type × with_parameters).
  - main added "moreau" as a third backend value with _MOREAU_OK skip handling.
  Resolved by combining both: backend = [cvxpy, qpax, moreau], plus the
  constraint_type and with_parameters axes from HEAD.

Adjacent fixups:
  - solve_budget extended to cover moreau (same JAX-array-reassembly
    cost profile as qpax: 3.5s ctcs / 4.5s nodal; cvxpy stays at 1.2s).
  - Comments generalized from QPAX-specific to "JAX-side backends" where
    they now apply to both qpax and moreau.

Second conflict (inside test_backend_float32_raises) was a mis-graft from
auto-merge — main's view ended test_backend right where HEAD's
test_backend_float32_raises begins, so git pasted main's test_backend
tail into the wrong function. Discarded the duplicate tail; kept HEAD's
pytest.raises wrapping.

+++ openscvx/discretization/discretize_linearize.py
@@ -188,7 +188,12 @@ class VectorizeDiscretizeLinearize(Discretizer):
                 in_axes=-1,
                 out_axes=-1,
             )
-            jacobians = pushforward(standard_basis)
+            # Cast to match primal's dtype.  standard_basis is built eagerly
+            # from jnp.eye (float32 when x64 is disabled), but jax.export
+            # traces with the literal dtype of the dummy arrays (float64 from
+            # np.ones), so primal may be float64 while the closure-captured
+            # standard_basis is float32.  jax.jvp requires identical dtypes.
+            jacobians = pushforward(standard_basis.astype(primal.dtype))
 
             A_d = jacobians[:, :, :, i0:i1]
             B_d = jacobians[:, :, :, i1:i2]
